Update chatgpt_client.py
# Extract code content only

Add BeautifulSoup to parse HTML (need for new feature)

Allow extraction of the content of code blocks only (not explanations or chat blabla)

All you need to do is fill in the argument: extract_code

# Improved interaction with the cat

Improved speed for copying and pasting prompts into the textarea (which used to take a very long time line by line).

Copying is now done in JS

Also, checking whether or not the script is generating is now done using the "Stop Generating" button. The old method didn't work (for me anyway).